feat: kde1d,kde2dのコマンドまわりを全面的に整理
Squashed commit of the following:
    feat: kde2dの主機能を,Commandsを用いて書き換え
    add: kde2d用のコマンドとパラメータ解釈器
    feat: kde1dの主機能を,Commandsを用いて書き換え
    add: kde1d用のコマンドとパラメータ解釈器
