--- src/matsu/num/statistics/kdeapp/kde2d/command/rule/package-info.java
@@ -1,11 +1,0 @@
-/*
- * Copyright © 2026 Matsuura Y.
- * 
- * This software is released under the MIT License.
- * http://opensource.org/licenses/mit-license.php
- */
-
-/**
- * kde2dのオプションコマンドに関するルールを扱うパッケージ.
- */
-package matsu.num.statistics.kdeapp.kde2d.command.rule;
