--- test/matsu/num/statistics/kdeapp/kde1d/command/ConsoleParameterInterpreterTest.java
@@ -1,125 +1,0 @@
-/*
- * Copyright © 2026 Matsuura Y.
- * 
- * This software is released under the MIT License.
- * http://opensource.org/licenses/mit-license.php
- */
-
-package matsu.num.statistics.kdeapp.kde1d.command;
-
-import static matsu.num.statistics.kdeapp.kde1d.command.ArgumentRequiringCommand.*;
-import static matsu.num.statistics.kdeapp.kde1d.command.NoArgumentCommand.*;
-import static matsu.num.statistics.kdeapp.kde1d.command.rule.CommandAssignmentRule.*;
-import static org.hamcrest.MatcherAssert.*;
-import static org.hamcrest.Matchers.*;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.experimental.runners.Enclosed;
-import org.junit.experimental.theories.DataPoints;
-import org.junit.experimental.theories.Theories;
-import org.junit.experimental.theories.Theory;
-import org.junit.runner.RunWith;
-
-import matsu.num.statistics.kdeapp.exception.IllegalParameterException;
-
-/**
- * {@link ConsoleParameterInterpreter} のテスト.
- */
-@RunWith(Enclosed.class)
-final class ConsoleParameterInterpreterTest {
-
-    @RunWith(Theories.class)
-    public static class 解釈生成の正常系に関するテスト {
-
-        @DataPoints
-        public static List<String[]> argsList;
-
-        @BeforeClass
-        public static void before_引数リストの作成() {
-            argsList = new ArrayList<>();
-
-            argsList.add(new String[] {});
-            argsList.add(
-                    new String[] {
-                            INPUT_FILE_PATH.commandString(), "test.txt" });
-            argsList.add(
-                    new String[] {
-                            INPUT_FILE_PATH.commandString(), "test.txt",
-                            ECHO_OFF.commandString() });
-            argsList.add(
-                    new String[] {
-                            ECHO_OFF.commandString(),
-                            INPUT_FILE_PATH.commandString(), "test.txt" });
-            argsList.add(
-                    new String[] {
-                            INPUT_FILE_PATH.commandString(), "test.txt",
-                            SEPARATOR.commandString(), "\t" });
-        }
-
-        @Theory
-        public void test_正常系の網羅テスト(String[] args) {
-            // 例外がスローされなければOK
-            ConsoleParameterInterpreter.from(args, nullRule());
-        }
-    }
-
-    public static class 解釈生成の異常系に関するテスト {
-
-        @Test(expected = IllegalParameterException.class)
-        public void test_パラメータに重複がある場合は例外_引数有り() {
-
-            String[] args = {
-                    INPUT_FILE_PATH.commandString(), "test.txt",
-                    INPUT_FILE_PATH.commandString(), "test.txt" };
-            ConsoleParameterInterpreter.from(args, nullRule());
-        }
-
-        @Test(expected = IllegalParameterException.class)
-        public void test_パラメータに重複がある場合は例外_引数無し() {
-
-            String[] args = { ECHO_OFF.commandString(), ECHO_OFF.commandString() };
-            ConsoleParameterInterpreter.from(args, nullRule());
-        }
-    }
-
-    public static class 解釈結果の取得に関するテスト {
-
-        /*
-         * インプットファイル と dummy-no-arg を設定
-         */
-
-        private final String file = "test.txt";
-
-        private ConsoleParameterInterpreter interpretation;
-
-        @Before
-        public void before_解釈の構築() {
-            // インプットファイル と dummy-no-arg を設定
-            String[] args = {
-                    INPUT_FILE_PATH.commandString(), file, ECHO_OFF.commandString()
-            };
-            interpretation = ConsoleParameterInterpreter.from(args, nullRule());
-        }
-
-        @Test
-        public void test_inputFileは設定ずみ() {
-            assertThat(interpretation.valueOf(INPUT_FILE_PATH).get(), is(file));
-        }
-
-        @Test
-        public void test_commentCharは設定されていない() {
-            assertThat(interpretation.valueOf(COMMENT_CHAR), is(Optional.empty()));
-        }
-
-        @Test
-        public void test_echoOffは設定済み() {
-            assertThat(interpretation.contains(ECHO_OFF), is(true));
-        }
-    }
-}
